changed parser

+++ src/controllers/task.controller.ts
@@ -1,10 +1,18 @@
 import { createTask, deleteById, getTaskById, updateById,getAllTask, archivedTask } from "../service/CRUD.js"
 import createHttpError from "http-errors"
 import { parsePaginationParams } from "../utils/parsePagination.js"
-import { parseFiltersFields } from "../utils/filterParsers.js"
+import { FilerFields, parseFiltersFields } from "../utils/filterParsers.js"
 import { parsedSort } from "../utils/parseSort.js"
+import type { Request, Response } from "express"
+import { AuthenticateUser } from "../middlewares/authMiddlewares.js"
+import { Task, UpdatedTaskDTO } from "../types/task.js"
 
-export const createTaskController = async (req, res) => {
+export const createTaskController = async (req: Request<{}, {}, UpdatedTaskDTO>, res: Response) => {
+    if (!req.user) {
+        return res.status(401).json({
+            message: 'Unauthorized'
+        })
+    }
     const owner = req.user._id
     const updateData = req.body
     const task = await createTask({
@@ -18,10 +26,18 @@ export const createTaskController = async (req, res) => {
     })
 }
 
-export const getAllTasksController = async (req, res) => {
+export const getAllTasksController = async (req: Request<{}, {}, Task, FilerFields>, res: Response) => {
+        if (!req.user) {
+        return res.status(401).json({
+            message: 'Unauthorized'
+        })
+    }
     const { page, perPage } = parsePaginationParams(req.query)
     const filter = parseFiltersFields(req.query)
-    const {sortOrder, sortBy,search} = parsedSort(req.query)
+    const { sortOrder, sortBy, search } = parsedSort(req.query)
+    if (!search) {
+        return undefined
+    }
     const owner = req.user._id
     const response = await getAllTask({
         page,
@@ -39,7 +55,23 @@ export const getAllTasksController = async (req, res) => {
     })
 }
 
-export const getTaskByIdController = async (req, res) => {
+export type TaskParams = {
+    taskId: string
+}
+declare global {
+    namespace Express {
+        interface Request {
+            user?: AuthenticateUser
+        }
+    }
+}
+export {}
+export const getTaskByIdController = async (req: Request<TaskParams, {}, {}>, res: Response) => {
+    if (!req.user) {
+        return res.status(401).json({
+            message: 'Unauthorized'
+        })
+    }
     const { taskId } = req.params
     const owner = req.user._id
     const response = await getTaskById({
@@ -55,10 +87,15 @@ export const getTaskByIdController = async (req, res) => {
     })
 }
 
-export const updateByIdController = async (req, res) => {
+export const updateByIdController = async (req: Request<TaskParams, {}, UpdatedTaskDTO>, res: Response) => {
+      if (!req.user) {
+    return res.status(401).json({
+      message: "Unauthorized",
+    });
+  }
     const { taskId } = req.params
+        const updateData = req.body
     const owner = req.user._id
-    const updateData = req.body
     const response = await updateById({ taskId, owner, updateData })
     if (!response) {
         throw createHttpError(404,'Task with current ID not found')
@@ -68,7 +105,12 @@ export const updateByIdController = async (req, res) => {
         data: response
     })
 } 
-export const deleteByIdController = async (req, res) => {
+export const deleteByIdController = async (req: Request<TaskParams, {}, Task>, res: Response) => {
+    if (!req.user) {
+        return res.status(401).json({
+            message: 'Unauthorized'
+        })
+    }
     const { taskId } = req.params
     const owner = req.user._id
     const response = await deleteById({taskId, owner})
@@ -77,7 +119,12 @@ export const deleteByIdController = async (req, res) => {
     }
     res.status(204).send()
 }
-export const archivedTaskController = async (req, res) => {
+export const archivedTaskController = async (req: Request<TaskParams, {}, Task>, res: Response) => {
+    if (!req.user) {
+        return res.status(401).json({
+            message: 'Unauthorized'
+        })
+    }
     const { taskId } = req.params
     const owner = req.user._id
     const response = await archivedTask({taskId, owner})
